Update index.html
Added id's to all of the <tr> elements, stating their time - this will be used in conjunction with the JS script to help create a function that loops through the current time and the times allocated to the <tr>'s.  Allowing me to change the background colour of these elements depending on whether or not they are in the Past, Present or Future.